+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- markdown output now tries to resemble the time entry calendar dialog
+
 ## [v0.53.1] - 2024-06-14
 
 ### Fixed